requirements.txt
# Translation and Language
ollama
xpinyin
transformers
torch
sentencepiece
accelerate

# TTS
piper-tts
pygame

# ENV
python-dotenv